Added a default error builder which the user can extend
